+++ src/plots/plots.R
@@ -741,3 +741,18 @@ addTitleToPlot <- function(myPlot, title) {
   }
   return(plotData[["popupContents"]])
 }
+
+fillDataWithQualityControlStatus <- function(data) {
+  status <- rep("NA", nrow(data))
+  status <- ifelse(data$anflag == 0, "Rejected", status)
+  status <- ifelse(data$active  > 0, "Active", status)
+  status <- ifelse(data$rejected > 0, "Rejected", status)
+  status <- ifelse(data$passive > 0, "Passive", status)
+  status <- ifelse(data$blacklisted > 0, "Blacklisted", status)
+  status <- ifelse(data$anflag  > 0, "Active(2)", status)
+  status <- ifelse(data$anflag == 4, "Rejected", status)
+  status <- ifelse(data$anflag == 8, "Blacklisted", status)
+  data$status <- status
+  data <- data[rev(order(data$status)),]
+  return(data)
+}