+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 ### Added
+- New "Number of Active Observations" plot type
 - Editable colour scale and data range on applicable plots
 - Possibility to choose units for physical quantities where applicable
 - Button to toggle "show only standard levels" (when applicable)